Separate trainer creation and training execution logic
